refactor: 💡 beautify the code
