+++ src/controllers/user/index.ts
@@ -1,3 +1,1 @@
-import routes from './routes';
-
-export default routes;
+export { default as userRoutes } from './routes';
