+++ src/libs/index.ts
@@ -1,9 +1,15 @@
-export { default as errorHandler } from './errorHandler';
-export { notFoundRoute, authorizationHandler, PermissionType } from './routes';
-export { default as validationHandler } from './validationHandler';
-export { default as sendResponse } from './ResponseHelper';
 export { default as Database } from './Database';
-export { IRequest } from './routes';
+export { PermissionType } from './constants';
+
+export {
+  notFoundRoute,
+  authorizationHandler,
+  validationHandler,
+  errorHandler,
+  sendResponse,
+  IRequest,
+  IError
+} from './routes';
 
 
 // TODO:  use single syntax for import/export
