+++ src/libs/routes/IErrorHandlerResponse.ts
@@ -1,4 +1,4 @@
-import IResponse from '../IResponse';
+import IResponse from './IResponse';
 
 export default interface IErrorHandlerResponse extends IResponse {
   error: string;
