+++ src/repositories/index.ts
@@ -1,4 +1,1 @@
-export { UserRepository } from './user';
-
-// import { UserRepository } from './user';
-// export { UserRepository };
+export { UserRepository, IUserModel } from './user';
